backend/requirements.txt
fastapi
uvicorn
asyncio
python-dotenv
manim
moviepy
openai
pydub
pydantic
anthropic
lmnt
groq
boto3
